feat: enhance delete functionality for anime with confirmation and reload

+++ ut8-angular/11-httpclient/src/app/features/animes/animes.ts
@@ -12,7 +12,12 @@ export class Animes {
   protected readonly animesApi: AnimeApi = inject(AnimeApi);
 
   protected deleteAnimeById(id: number) {
+    if (!confirm('¿Seguro que quieres eliminar este anime?')) return;
     this.animesApi.deleteAnimeById(id).subscribe({
+      next: () => {
+        this.animesApi.getAllAnimes.reload();
+        console.log('Anime eliminado correctamente');
+      },
       error: (err) => {
         console.error('Error al eliminar el anime', err);
       },
